Add "AbsorbChip" into "gadgets" dir
